Refactor platform dependent information getter logic
This refactors the platform dependent information getter logic to
childclasss. This makes the code much cleaner but also avoid unneeded
cpu cycles for the *if halium* checks everywhere.

This is prepares for a *cached* platfrom childclass that will be able
to provide information to confined applications without needing to resort
to getprop or dts tree access. This makes it also possible to restrict
apps access to certain information.