Move to yaml for config files
This moves to a more elegant solution for device info's config files,
this makes it easier to understand but also easier to overlay as we
don't have the need for the alias file as alias names are defined in the
config file for the device.

This also builds upon a way for users to override values where
deviceinfo can generate and overlay a user defined yaml file.

The way the overlaying logic work here is that we first take the device
specific config into a variable, we then "merge" values with default
values, this will only add the values from default that is not set by
device config.

Config files also is made in a way that we can group devices with a
common vendor under the same file and allow to have default values
across all devices with the same vendor (in that file) but also allow
standalone configfiles.

Syntax
default:
  ... (values) ...

[codename]
  Names: [list of aliases]
  ... (values) ...